split transfer details tables into two components, added action bar component for transfers
